fix other ordering tests
